fix(Prefetcher.sh):fix nl_idx error fix(NextLinePrefetch.sh):delete redundant compare diff code
